# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        example: Validation failed
        type: string
      message:
        example: Email is required
        type: string
    type: object
...
  dto.UserProfile:
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
    type: object
  dto.UserStats:
    properties:
      monthly_expenses:
        example: 450.75
        type: number
      total_budgets:
        example: 3
        type: integer
      total_categories:
        example: 5
        type: integer
      total_expenses:
        example: 1250.5
        type: number
      weekly_expenses:
        example: 125.25
        type: number
    type: object
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Система управления личными финансами
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Finance API
  version: "1.0"
paths:
...
  /user/account:
    delete:
      consumes:
      - application/json
      description: Полное удаление аккаунта пользователя и всех связанных данных
      produces:
      - application/json
      responses:
        "200":
          description: Аккаунт успешно удален
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление аккаунта пользователя
      tags:
      - User
  /user/profile:
    get:
      consumes:
      - application/json
      description: Получение информации о текущем пользователе
      produces:
      - application/json
      responses:
        "200":
          description: Профиль пользователя
          schema:
            $ref: '#/definitions/dto.UserProfile'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение профиля пользователя
      tags:
      - User
  /user/stats:
    get:
      consumes:
      - application/json
      description: Получение общей статистики по расходам, категориям и бюджетам
      produces:
      - application/json
      responses:
        "200":
          description: Статистика пользователя
          schema:
            $ref: '#/definitions/dto.UserStats'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение статистики пользователя
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"